refactor `Team` into `types` module
