+++ berserk/__init__.py
@@ -10,7 +10,8 @@ __email__ = berserk_metadata["Author-email"]
 __version__ = berserk_metadata["Version"]
 
 
-from .clients import Client, Team
+from .clients import Client
+from .types import Team
 from .session import TokenSession
 from .session import Requestor
 from .enums import PerfType
