+++ berserk/types/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+import requests
+
+from .team import Team
+
+__all__ = [
+    "Team",
+]
